+++ CHANGELOG.md
@@ -4,5 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [2.0.0] - 2021-11-26
+- Changed the project name to "Phantom" and URL to https://github.com/EONRaider/BCA-Phantom
+- Changed the CLI parsing to accept a full URL for the server instead of different parameters for hostname and port number
+- Greatly simplified the "client" and "build" modules by re-implementing them without using "configparser" for client configuration
+- Added HTTP mode for both servers and clients
+- Updated distribution binaries for Linux and Windows
+
 ## [1.0.0] - 2021-11-19
 - Public release